Redo file structure to move logic to services
