+++ src/main/java/com/foodforfriends/web/restaurant/RestaurantService.java
@@ -1,0 +1,5 @@
+package com.foodforfriends.web.restaurant;
+
+public class RestaurantService {
+
+}
