+++ src/main/java/com/foodforfriends/web/review/ReviewController.java
@@ -1,4 +1,4 @@
-package com.foodforfriends.web;
+package com.foodforfriends.web.review;
 
 import java.net.URI;
 import java.net.URISyntaxException;
