+++ src/main/java/com/foodforfriends/web/review/ReviewService.java
@@ -1,0 +1,5 @@
+package com.foodforfriends.web.review;
+
+public class ReviewService {
+
+}
