+++ src/main/java/com/foodforfriends/web/user/UserController.java
@@ -1,4 +1,4 @@
-package com.foodforfriends.web;
+package com.foodforfriends.web.user;
 
 import java.net.URI;
 import java.net.URISyntaxException;
