+++ src/main/java/com/foodforfriends/web/user/UserService.java
@@ -1,0 +1,5 @@
+package com.foodforfriends.web.user;
+
+public class UserService {
+
+}
